requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
langchain==0.3.9
langchain-google-genai==2.0.5
langgraph==0.2.50
langchain-core==0.3.21
